Refactor: retire unused 16/17 region control and canonicalize DRCT
Keep the delegated-region codec and terminal table under
comm_provider_control. Remove native allocate/release methods while
leaving control_payload unchanged, and fold frozen goldens into
canonical tests.
