+++ src/common/hierarchical/worker.h
@@ -126,13 +126,6 @@ public:
     void control_comm_init(int worker_id, const std::string &request_shm_name) {
         manager_.control_comm_init(worker_id, request_shm_name.c_str());
     }
-    void
-    control_region_allocate(int worker_id, const std::string &request_shm_name, const std::string &reply_shm_name) {
-        manager_.control_region_allocate(worker_id, request_shm_name.c_str(), reply_shm_name.c_str());
-    }
-    void control_region_release(int worker_id, const std::string &request_shm_name, const std::string &reply_shm_name) {
-        manager_.control_region_release(worker_id, request_shm_name.c_str(), reply_shm_name.c_str());
-    }
 
     ControlResult
     control_digest_only(WorkerType type, int worker_id, uint64_t sub_cmd, const uint8_t *digest, double timeout_s) {
